refactor: extract utility methods in EvmUtils
